Closes #8, A List Apart RSS feed
