+++ package.json
@@ -5,8 +5,6 @@
   "private": true,
   "main": "app.js",
   "scripts": {
-    "preinstall": "npm install -g bower gulp http-perf",
-    "postinstall": "bower install",
     "prestart": "npm install",
     "start": "node ./bin/www",
     "test": "gulp perf"
@@ -29,16 +27,21 @@
   "homepage": "https://github.com/mefernandez/temply",
   "dependencies": {
     "body-parser": "~1.13.2",
+    "cheerio": "^0.19.0",
     "connect-livereload": "^0.5.3",
     "cookie-parser": "~1.3.5",
     "debug": "~2.2.0",
     "express": "~4.13.1",
     "jade": "~1.11.0",
+    "lodash": "^3.10.1",
     "morgan": "~1.6.1",
     "serve-favicon": "~2.3.0",
-    "temply-express": "file:modules/temply-express"
+    "superagent": "^1.4.0",
+    "temply-express": "file:modules/temply-express",
+    "xml2js": "^0.4.12"
   },
   "devDependencies": {
+    "chai": "^3.3.0",
     "gulp": "^3.9.0",
     "gulp-express": "^0.3.5",
     "gulp-live-server": "0.0.28",
