+++ routes/index.js
@@ -4,7 +4,7 @@ var router = express.Router();
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  console.log('routes');
+  
   res.render('index.html');
 });
