--- template/cms-data-rss-feed.js
@@ -1,5 +1,0 @@
-var request = require('superagent');
-
-exports.execute = function(callback) {
-
-}
