+++ template/index.html
@@ -114,18 +114,18 @@
       <section id="rss-live-demo" class="row">
         <div class="col-md-6">
           <h2>RSS Plugin</h2>
-          <p>This content is served lived from <a href="http://feeds.feedburner.com/alistapart/main">A List Apart RSS feed</a></p>
-          <div id="ala-carousel" class="owl-carousel cms-data-rss-feed cms-render-article">
-            <article>
-              <h4>Article's title</h4>
+          <p>This content is served live from <a href="http://feeds.feedburner.com/alistapart/main">A List Apart RSS feed</a></p>
+          <div id="ala-carousel" class="owl-carousel cms-data-rss-feed">
+            <article class="cms-render-rss-feed">
+              <a href="#"><h4>Article's title</h4></a>
               <p>Article's content, Lorem ipsum dolor sit amet, consectetur adipisicing elit. Sunt dicta totam nemo explicabo asperiores officiis cupiditate dolorem, quibusdam iste dolorum quaerat delectus architecto et eos! Porro omnis ad molestiae blanditiis.</p>
             </article>
-            <article>
-              <h4>Article's title</h4>
+            <article class="cms-render-rss-feed">
+              <a href="#"><h4>Article's title</h4></a>
               <p>Article's content, Lorem ipsum dolor sit amet, consectetur adipisicing elit. Sunt dicta totam nemo explicabo asperiores officiis cupiditate dolorem, quibusdam iste dolorum quaerat delectus architecto et eos! Porro omnis ad molestiae blanditiis.</p>
             </article>
-            <article>
-              <h4>Article's title</h4>
+            <article class="cms-render-rss-feed">
+              <a href="#"><h4>Article's title</h4></a>
               <p>Article's content, Lorem ipsum dolor sit amet, consectetur adipisicing elit. Sunt dicta totam nemo explicabo asperiores officiis cupiditate dolorem, quibusdam iste dolorum quaerat delectus architecto et eos! Porro omnis ad molestiae blanditiis.</p>
             </article>
           </div>
